Move functions to proper files according to their objective
